# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IoT Trap Monitoring System API",
        "description": "API для системы мониторинга IoT ловушек для борьбы с вредителями. Включает управление шлюзами, ловушками и регистрацию событий срабатывания.",
        "contact": {
            "email": "support@trap-monitoring.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:3000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.trap-monitoring.local/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/v1/facilities": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "List facilities",
                "operationId": "getFacilities",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "warehouse",
                                "restaurant",
                                "office",
                                "retail",
                                "industrial"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FacilityResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Create facility",
                "operationId": "createFacility",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFacilityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FacilityResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/facilities/{id}": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get facility details",
                "operationId": "getFacilityById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FacilityWithGatewaysResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Update facility",
                "operationId": "updateFacility",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFacilityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FacilityResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Delete facility",
                "operationId": "deleteFacility",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/v1/facilities/{id}/gateways": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get facility gateways",
                "operationId": "getFacilityGateways",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GatewayResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateFacilityRequest": {
                "required": [
                    "name",
                    "address",
                    "contact_person",
                    "contact_phone",
                    "type",
                    "area"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Main Warehouse"
                    },
                    "address": {
                        "required": [
                            "street",
                            "city",
                            "postal_code",
                            "country"
                        ],
                        "properties": {
                            "street": {
                                "type": "string",
                                "example": "123 Main Street"
                            },
                            "city": {
                                "type": "string",
                                "example": "Kyiv"
                            },
                            "postal_code": {
                                "type": "string",
                                "example": "01001"
                            },
                            "country": {
                                "type": "string",
                                "example": "Ukraine"
                            }
                        },
                        "type": "object"
                    },
                    "contact_person": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "contact_phone": {
                        "type": "string",
                        "example": "+380501234567"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "warehouse",
                            "restaurant",
                            "office",
                            "retail",
                            "industrial"
                        ],
                        "example": "warehouse"
                    },
                    "area": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1500.5
                    },
                    "description": {
                        "type": "string",
                        "example": "Main distribution center",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active",
                        "nullable": true
                    },
                    "customer_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "01963e34-e7bc-7890-b23c-000000000001",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateFacilityRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Main Warehouse"
                    },
                    "address": {
                        "required": [
                            "street",
                            "city",
                            "postal_code",
                            "country"
                        ],
                        "properties": {
                            "street": {
                                "type": "string",
                                "example": "123 Main Street"
                            },
                            "city": {
                                "type": "string",
                                "example": "Kyiv"
                            },
                            "postal_code": {
                                "type": "string",
                                "example": "01001"
                            },
                            "country": {
                                "type": "string",
                                "example": "Ukraine"
                            }
                        },
                        "type": "object"
                    },
                    "contact_person": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "contact_phone": {
                        "type": "string",
                        "example": "+380501234567"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "warehouse",
                            "restaurant",
                            "office",
                            "retail",
                            "industrial"
                        ],
                        "example": "warehouse"
                    },
                    "area": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 1500.5
                    },
                    "description": {
                        "type": "string",
                        "example": "Main distribution center",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
            "FacilityResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "properties": {
                            "street": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            },
                            "postal_code": {
                                "type": "string"
                            },
                            "country": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "contact_person": {
                        "type": "string"
                    },
                    "contact_phone": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "area": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "FacilityWithGatewaysResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "properties": {
                            "street": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            },
                            "postal_code": {
                                "type": "string"
                            },
                            "country": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "contact_person": {
                        "type": "string"
                    },
                    "contact_phone": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "area": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "gateways": {
                        "description": "Шлюзы, установленные на объекте",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GatewayResource"
                        }
                    },
                    "gateway_count": {
                        "description": "Количество шлюзов",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "GatewayResource": {
                "properties": {
                    "id": {
                        "description": "UUID шлюза",
                        "type": "string",
                        "format": "uuid",
                        "example": "9d4e1c45-6b2a-4f8e-9c7d-3a1b5e8f4d2c"
                    },
                    "hardware_id": {
                        "description": "Идентификатор оборудования",
                        "type": "string",
                        "example": "GW-001-ABC123"
                    },
                    "facility_id": {
                        "description": "UUID объекта",
                        "type": "string",
                        "format": "uuid",
                        "example": "019bac36-54b6-70dc-94c3-cea624a51791"
                    },
                    "name": {
                        "description": "Название шлюза",
                        "type": "string",
                        "example": "Gateway Floor 1"
                    },
                    "location": {
                        "description": "Координаты расположения",
                        "properties": {
                            "latitude": {
                                "type": "number",
                                "format": "float",
                                "example": 50.4501
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float",
                                "example": 30.5234
                            }
                        },
                        "type": "object"
                    },
                    "status": {
                        "description": "Статус шлюза",
                        "type": "string",
                        "enum": [
                            "online",
                            "offline",
                            "maintenance"
                        ],
                        "example": "online"
                    },
                    "is_online": {
                        "description": "Онлайн статус",
                        "type": "boolean",
                        "example": true
                    },
                    "last_seen_at": {
                        "description": "Время последней активности",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00+00:00"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T10:00:00+00:00"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00+00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "HmacSignature": {
                "type": "apiKey",
                "description": "HMAC SHA-256 signature для защиты IoT endpoints. Используйте hash_hmac(\"sha256\", payload, gateway_api_key)",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}